Complited dataproviders. Some refactoring.

+++ src/main/java/ru/sfedu/tavern/dataproviders/JdbcAPI.java
@@ -12,6 +12,10 @@ import ru.sfedu.tavern.model.Result;
 import ru.sfedu.tavern.model.StatusType;
 import ru.sfedu.tavern.model.entities.ClassType;
 import ru.sfedu.tavern.model.entities.Entity;
+import ru.sfedu.tavern.model.entities.Message;
+import ru.sfedu.tavern.model.entities.OurUser;
+import ru.sfedu.tavern.model.entities.Platform;
+import ru.sfedu.tavern.model.entities.PlatformUser;
 
 /**
  *
@@ -33,7 +37,7 @@ public class JdbcAPI implements IDataProvider{
     }
 
     @Override
-    public Result insert(ArrayList<Entity> object) throws Exception {
+    public Result insert(List<Entity> object) throws Exception {
         ClassType classType = object.get(0).getClassType();
         try {
             object.stream().forEach(iterator -> {
@@ -47,25 +51,49 @@ public class JdbcAPI implements IDataProvider{
     }
 
     @Override
-    public Result insert(Entity object) throws Exception {
+    public Result insert(Optional<Entity> optObject) throws Exception {
+        if(!optObject.isPresent()) return new Result(StatusType.ERROR);
+        Entity object = optObject.get();
         ArrayList<Entity> objList = new ArrayList();
         objList.add(object);
         return insert(objList);
     }
 
     @Override
-    public Result update(Entity object) throws Exception {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    public Result update(Optional<Entity> optObject) throws Exception {
+        if(!optObject.isPresent()) return new Result(StatusType.ERROR);
+        Entity object = optObject.get();
+        ClassType classType = object.getClassType();
+        try {
+            instance.execQuery("UPDATE " + classType.getTableName() + " " + setExpGen(object) + " WHERE id=" + object.getId() + ";", classType);   
+        } catch (Exception ex) {
+            return new Result(StatusType.ERROR);
+        }
+        return new Result(StatusType.OK);
     }
 
     @Override
-    public Result delete(Entity object) throws Exception {
+    public Result delete(Optional<Entity> optObject) throws Exception {
+        if(!optObject.isPresent()) return new Result(StatusType.ERROR);
+        Entity object = optObject.get();
         ClassType classType = object.getClassType();
         try {
             instance.execQuery("DELETE FROM " + classType.getTableName() + " WHERE id=" + object.getId() + ";", classType);   
         } catch (Exception ex) {
             return new Result(StatusType.ERROR);
         }
+        return new Result(StatusType.OK);
+    }
+
+    @Override
+    public Result delete(List<Entity> objects) throws Exception {
+        objects.stream().forEach(e -> {
+            try{
+                delete(Optional.ofNullable(e));
+            } catch (Exception ex) {
+                logger.warn(ex);
+            }
+        });
         return new Result(StatusType.OK);
     }
 
@@ -99,4 +127,55 @@ public class JdbcAPI implements IDataProvider{
         return result;
     }
 
+    private String setExpGen(Entity obj) {
+        String res = "";
+        switch(obj.getClassType()) {
+            case MESSAGE:
+                Message mes = (Message) obj;
+                res = " SET id=" + mes.getId() + 
+                      ",senderId=" + mes.getSenderId() +
+                      ",text='" + mes.getText() + '\'' +
+                      ",sendTime='" + mes.getSendTime() + '\'' +
+                      ",platformId=" + mes.getPlatformId();
+                break;
+            case OURUSER:
+                OurUser ou = (OurUser) obj;
+                res = " SET id=" + ou.getId() + 
+                      ",login='" + ou.getLogin() + '\'' +
+                      ",passwordHash='" + ou.getPasswordHash() + '\'' +
+                      ",email='" + ou.getEmail() + '\'' +
+                      ",lastAct='" + ou.getLastAct() + '\'';
+                break;
+            case PLATFORM:
+                Platform pl = (Platform) obj;
+                res = " SET id=" + pl.getId() + 
+                      ",domain='" + pl.getDomain() + '\'' +
+                      ",key='" + pl.getKey() + '\'' +
+                      ",ownerId=" + pl.getOwnerId();
+                break;
+            case PLATFORMUSER:
+                PlatformUser pu = (PlatformUser) obj;
+                res = " SET id=" + pu.getId() + 
+                      ",login='" + pu.getLogin() + '\'' +
+                      ",avatarLink='" + pu.getAvatarLink() + '\'' +
+                      ",lastAct='" + pu.getLastAct() + '\'' +
+                      ",banned=" + pu.getBanned() +
+                      ",platformId=" + pu.getPlatformId();
+                break;
+        }
+        return res;
+    }
+    
+    @Override
+    public long getNextId(ClassType type) {
+        List<Entity> list = new ArrayList();
+        try {
+            list.addAll(select(type));
+        } catch(Exception ex) {
+            return 1;
+        }
+        long maxId = list.stream().max((p1, p2) -> Long.valueOf(p1.getId()).compareTo(p2.getId())).get().getId();
+        return maxId + 1;
+    }
+    
 }
